+++ crates/python/qcs_sdk/qpu/quilc.pyi
@@ -9,11 +9,11 @@ class CompilerOpts:
     """A set of options that determine the behavior of compiling programs with quilc."""
 
     @property
-    def timeout(self) -> Optional[int]:
+    def timeout(self) -> Optional[float]:
         """The number of seconds to wait before timing out. If `None`, there is no timeout."""
         ...
     def __new__(
-        cls, timeout: Optional[int] = DEFAULT_COMPILER_TIMEOUT
+        cls, timeout: Optional[float] = DEFAULT_COMPILER_TIMEOUT
     ) -> "CompilerOpts": ...
     @staticmethod
     def default() -> "CompilerOpts": ...
